requirements.txt
tensorflow
python-interface==1.5.3
numpy
numba